Package/requirements.txt
requests==2.18.4
Keras==2.2.0
numpy==1.15.0
matplotlib==2.2.3
nltk==3.3
python_Levenshtein==0.12.0
